rl/db/db_requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiohttp-cors==0.8.1
aiosignal==1.3.2
ale-py==0.11.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
colorful==0.5.6
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.1
distlib==0.3.9
dm-tree==0.1.9
einops==0.8.1
Farama-Notifications==0.0.4
fastapi==0.115.12
filelock==3.18.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.5.1
google-api-core==2.24.2
google-auth==2.40.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
gymnasium==1.1.1
h11==0.16.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.6
jiwer==3.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mazelib==0.9.16
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
mysqlclient==2.2.7
networkx==3.4.2
numpy==2.2.6
omegaconf==2.3.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pettingzoo==1.25.0
pillow==11.2.1
platformdirs==4.3.8
prometheus_client==0.22.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.0
psutil==7.0.0
py-spy==0.4.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.8
pydantic==2.11.5
pydantic_core==2.33.2
pygame==2.6.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
ray==2.46.0
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
sb3_contrib==2.6.0
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
stable_baselines3==2.6.0
starlette==0.46.2
SuperSuit==3.9.3
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
-e git+https://github.com/til-ai/til-25-environment.git@56389791f26baeeec2b82af81b445736e6b75665#egg=til_environment&subdirectory=..\..\til-25-environment
tinyscaler==1.2.8
torch==2.7.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
virtualenv==20.31.2
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.0
